Refactor WNN Core for numerical determinism and memory safety
Upgraded all physics state variables (RK4 integrator, Duffing Oscillator) to `long double` precision and implemented Kahan summation to prevent truncation errors. Transitioned concurrency to C++20 `std::jthread` and introduced strict clock-derived `dt` steps for deterministic execution. Hardened memory via `std::unique_ptr` for pcap sockets and introduced atomic flush mechanisms for the Vector B transduction queue. Finally, added Schmitt trigger hysteresis and EMA smoothing to the AILEE Trust Layer, and applied Hann window filtering to the FFT pre-processing.

+++ src/calibration.cpp
@@ -12,7 +12,7 @@
 namespace wave_native {
 namespace core {
 
-bool CalibrationMode::run_sweep(double& out_omega) {
+bool CalibrationMode::run_sweep(long double& out_omega) {
     std::cout << "[Calibration] Starting non-linear frequency sweep (500 Hz -> 20 kHz)...\n";
 
     double best_omega = 500.0;
@@ -45,9 +45,9 @@ bool CalibrationMode::run_sweep(double& out_omega) {
             // High energy -> higher confidence (bounded to [0,1])
             double energy = 0.5 * std::pow(state.x_dot, 2) + 0.5 * alpha * std::pow(state.A, 2) + 0.25 * beta * std::pow(state.A, 4);
             score.confidence_score = std::min(1.0, std::abs(energy) * 0.5);
-            score.consistency_score = std::min(1.0, 1.0 / (1.0 + std::abs(state.x_dot)));
+            score.consistency_score = std::min(1.0, static_cast<double>(1.0 / (1.0 + std::abs(state.x_dot))));
             score.safety_score = 1.0;
-            score.determinism_score = std::min(1.0, 1.0 / (1.0 + std::abs(state.A)));
+            score.determinism_score = std::min(1.0, static_cast<double>(1.0 / (1.0 + std::abs(state.A))));
 
             sum_aggregate += score.aggregate_score();
         }
@@ -94,7 +94,7 @@ bool CalibrationMode::save_profile(const std::string& path, double omega, double
     return true;
 }
 
-bool CalibrationMode::load_profile(const std::string& path, double& omega, double& alpha, double& beta, double& delta) {
+bool CalibrationMode::load_profile(const std::string& path, long double& omega, long double& alpha, long double& beta, long double& delta) {
     std::ifstream in(path, std::ios::binary);
     if (!in) return false;
     in.read(reinterpret_cast<char*>(&omega), sizeof(omega));
